Update tests to add more complexity.
